requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
sqlmodel==0.0.22
pydantic==2.10.3
yt-dlp>=2026.3.17
youtube-transcript-api>=1.2.4
sumy==0.11.0
transformers==4.46.3
torch>=2.2.0
sentencepiece==0.2.0
gTTS==2.5.4
langdetect==1.0.9
python-multipart==0.0.19
python-dotenv==1.0.1
groq==0.20.0
static-ffmpeg>=3.0
